добавляет загрузку фото и аватара
