+++ css/style.css
@@ -1059,3 +1059,7 @@ fieldset {
   font-size: 0.85em;
   color: #ff5635;
 }
+
+.foto-preview__hidden {
+  display: none;
+}
