+++ index.html
@@ -169,6 +169,7 @@
               <label for="images" class="ad-form__drop-zone">Загрузить<br>фото...</label>
             </div>
             <div class="ad-form__photo">
+              <img class="foto-preview foto-preview__hidden" src="" alt="Фото жилья пользователя" width="70" height="70">
             </div>
           </div>
         </fieldset>
